+++ R/globalVariables.R
@@ -1,0 +1,4 @@
+## Use to handle the CRAN check NOTE
+utils::globalVariables(
+  c("AMP", "DEL")
+)